# docs/specifications/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dog Walking Management API
  description: |
    The Stardogwalker management API — Cardiff, South Wales.

    This API is designed **API-first** to support multiple interaction
    channels: web app, native mobile app, conversational chat interfaces
    (e.g. WhatsApp bot), and agentic AI applications. All business logic
    lives here; no channel-specific logic is embedded in the API.

    **Design principles:**
    - Stateless — every request authenticates via JWT bearer token
    - Fine-grained actions — discrete endpoints for each state transition
      (e.g. `/accept`, `/decline`) to support unambiguous intent from any
      client type including AI agents
    - Consistent response envelopes — list responses always include
      `data[]` and `pagination`; errors always return `code`, `message`,
      and optional `details`
    - Versioned — breaking changes will be introduced under `/v2` without
      removing `/v1` until all clients have migrated

    Covers: interest registration (public), authentication, dog owner and
    walker management, dog profiles, walk requests, walk scheduling and
    execution, real-time walk updates, and invoicing.
  version: 1.0.0
  contact:
    name: Dog Walking API Support
    email: support@dogwalking.example.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/v1
    description: Local development server
  - url: https://api.dogwalking.example.com/v1
    description: Production server
...
security:
  - bearerAuth: []
...
paths:
...
  /owners:
    get:
      tags:
        - Owners
      summary: List dog owners
      description: Retrieve a paginated list of dog owners. Accessible by walkers.
      operationId: listOwners
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A list of dog owners
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerList'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Owners
      summary: Create a dog owner
      description: Create a new dog owner profile.
      operationId: createOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerRequest'
      responses:
        '201':
          description: Dog owner created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /owners/{ownerId}:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
    get:
      tags:
        - Owners
      summary: Get a dog owner
      description: Retrieve details of a specific dog owner.
      operationId: getOwner
      responses:
        '200':
          description: Dog owner details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Owners
      summary: Update a dog owner
      description: Update the profile of a specific dog owner.
      operationId: updateOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerRequest'
      responses:
        '200':
          description: Dog owner updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Owners
      summary: Delete a dog owner
      description: Delete a dog owner profile.
      operationId: deleteOwner
      responses:
        '204':
          description: Dog owner deleted
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Dog Walkers
  # ---------------------------------------------------------------------------
...
  /owners/{ownerId}/dogs:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
    get:
      tags:
        - Dogs
      summary: List dogs for an owner
      description: Retrieve all dogs belonging to a specific owner.
      operationId: listDogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A list of dogs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DogList'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Dogs
      summary: Add a dog to an owner
      description: Create a new dog record associated with a specific owner.
      operationId: createDog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DogRequest'
      responses:
        '201':
          description: Dog created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /walks:
    get:
      tags:
        - Walks
      summary: List walks
      description: |
        Retrieve a paginated list of scheduled or completed walks. Dog owners
        see walks for their dogs; walkers see their assigned walks.
      operationId: listWalks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter walks by status
          schema:
            $ref: '#/components/schemas/WalkStatus'
        - name: date
          in: query
          description: Filter walks by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of walks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkList'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /walks/{walkId}:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    get:
      tags:
        - Walks
      summary: Get a walk
      description: Retrieve details of a specific walk.
      operationId: getWalk
      responses:
        '200':
          description: Walk details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Walks
      summary: Update a walk
      description: Update details of a scheduled walk (e.g. start time, route notes).
      operationId: updateWalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkEdit'
      responses:
        '200':
          description: Walk updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /walks/{walkId}/cancel:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    post:
      tags:
        - Walks
      summary: Cancel a walk
      description: Cancel a scheduled walk.
      operationId: cancelWalk
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkCancel'
      responses:
        '200':
          description: Walk cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Walk Updates (images and notes)
  # ---------------------------------------------------------------------------
...
  /recurring-walks:
    get:
      tags:
        - Recurring Walks
      summary: List recurring walks
      description: Retrieve all recurring walk schedules for the authenticated owner, or all schedules for the walker.
      operationId: listRecurringWalks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter by recurring walk status
          schema:
            $ref: '#/components/schemas/RecurringWalkStatus'
      responses:
        '200':
          description: A list of recurring walks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalkList'
        '401':
          $ref: '#/components/responses/Unauthorised'

    post:
      tags:
        - Recurring Walks
      summary: Create a recurring walk
      description: Set up a recurring walk schedule. Generates WalkRequests automatically on each occurrence.
      operationId: createRecurringWalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringWalkCreate'
      responses:
        '201':
          description: Recurring walk created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /recurring-walks/{recurringWalkId}:
    parameters:
      - $ref: '#/components/parameters/RecurringWalkId'
    get:
      tags:
        - Recurring Walks
      summary: Get a recurring walk
      description: Retrieve details of a specific recurring walk schedule.
      operationId: getRecurringWalk
      responses:
        '200':
          description: Recurring walk details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalk'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Recurring Walks
      summary: Update a recurring walk
      description: Update the schedule details of an active recurring walk.
      operationId: updateRecurringWalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringWalkUpdate'
      responses:
        '200':
          description: Recurring walk updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /recurring-walks/{recurringWalkId}/pause:
    parameters:
      - $ref: '#/components/parameters/RecurringWalkId'
    post:
      tags:
        - Recurring Walks
      summary: Pause a recurring walk
      description: Pause an active recurring walk. No new WalkRequests are generated while paused.
      operationId: pauseRecurringWalk
      responses:
        '200':
          description: Recurring walk paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /recurring-walks/{recurringWalkId}/resume:
    parameters:
      - $ref: '#/components/parameters/RecurringWalkId'
    post:
      tags:
        - Recurring Walks
      summary: Resume a recurring walk
      description: Resume a paused recurring walk. Request generation resumes from the next scheduled occurrence.
      operationId: resumeRecurringWalk
      responses:
        '200':
          description: Recurring walk resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /recurring-walks/{recurringWalkId}/cancel:
    parameters:
      - $ref: '#/components/parameters/RecurringWalkId'
    post:
      tags:
        - Recurring Walks
      summary: Cancel a recurring walk
      description: Permanently cancel a recurring walk. Stops all future generation of WalkRequests. Any pending generated requests are cancelled automatically.
      operationId: cancelRecurringWalk
      responses:
        '200':
          description: Recurring walk cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringWalk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Invoices
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    OwnerId:
      name: ownerId
      in: path
      required: true
      description: Unique identifier of the dog owner
      schema:
        type: string
        format: uuid
...
    WalkId:
      name: walkId
      in: path
      required: true
      description: Unique identifier of the walk
      schema:
        type: string
        format: uuid
...
    RecurringWalkId:
      name: recurringWalkId
      in: path
      required: true
      description: Unique identifier of the recurring walk
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validationError:
              summary: Validation error
              value:
                code: VALIDATION_ERROR
                message: One or more fields failed validation.
                details:
                  - field: email
                    issue: Must be a valid email address.
                  - field: durationMinutes
                    issue: Must be between 15 and 240.
            noMatchingRate:
              summary: No matching rate found
              value:
                code: NO_MATCHING_RATE
                message: No rate found for this walk type, dog count, and duration. Please update the walker's rate card.
            invalidStateTransition:
              summary: Invalid state transition
              value:
                code: INVALID_STATE_TRANSITION
                message: This action cannot be performed on a resource in its current status.
            schedulingConflict:
              summary: Scheduling conflict
              value:
                code: SCHEDULING_CONFLICT
                message: The walker already has an overlapping walk scheduled for this time.
    Unauthorised:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              summary: Missing bearer token
              value:
                code: AUTHENTICATION_REQUIRED
                message: A valid bearer token is required to access this resource.
            invalidCredentials:
              summary: Invalid login credentials
              value:
                code: INVALID_CREDENTIALS
                message: The email or password provided is incorrect.
            tokenExpired:
              summary: Expired token
              value:
                code: TOKEN_EXPIRED
                message: The access token has expired. Please refresh your session.
    Forbidden:
      description: Authenticated user does not have permission to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            insufficientRole:
              summary: Insufficient role
              value:
                code: FORBIDDEN
                message: Your role does not have permission to perform this action.
            notOwnResource:
              summary: Not own resource
              value:
                code: FORBIDDEN
                message: You can only access your own records.
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                code: RESOURCE_NOT_FOUND
                message: The requested resource could not be found.
...
  schemas:
...
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - total
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          description: Total number of matching records
          example: 42
...
    OwnerRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phone
      properties:
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Smith
        email:
          type: string
          format: email
          example: jane.smith@example.com
        phone:
          type: string
          example: +44 7700 900123
        address:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
          description: Additional notes about this owner
          example: Prefers morning walks.
    Owner:
      allOf:
        - $ref: '#/components/schemas/OwnerRequest'
        - type: object
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
              example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
            createdAt:
              type: string
              format: date-time
              example: '2025-01-15T09:00:00Z'
            updatedAt:
              type: string
              format: date-time
              example: '2025-01-15T09:00:00Z'
    OwnerList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Owner'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Walkers --------------------------------------------------------------
...
    DogRequest:
      type: object
      required:
        - name
        - breed
        - dateOfBirth
      properties:
        name:
          type: string
          example: Max
        breed:
          type: string
          example: Labrador Retriever
        dateOfBirth:
          type: string
          format: date
          example: '2020-06-15'
        sex:
          type: string
          enum:
            - male
            - female
          example: male
        colour:
          type: string
          example: Golden
        microchipNumber:
          type: string
          description: ISO 11784/11785 microchip identifier
          example: '985141000123456'
        vaccinations:
          type: array
          items:
            $ref: '#/components/schemas/Vaccination'
        vetContact:
          $ref: '#/components/schemas/VetContact'
        medicalNotes:
          type: string
          description: Any medical conditions or special care instructions
          example: Allergic to certain grasses. Requires Apoquel daily.
        behaviourNotes:
          type: string
          description: Behavioural notes relevant for walkers
          example: Can be nervous around large dogs; keep on lead near traffic.
    Dog:
      allOf:
        - $ref: '#/components/schemas/DogRequest'
        - type: object
          required:
            - id
            - ownerId
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
              example: c1a2b3d4-5678-90ab-cdef-1234567890ab
            ownerId:
              type: string
              format: uuid
              example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
            createdAt:
              type: string
              format: date-time
              example: '2025-02-01T10:30:00Z'
            updatedAt:
              type: string
              format: date-time
              example: '2025-02-01T10:30:00Z'
    DogList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dog'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Vaccination:
      type: object
      required:
        - name
        - dateAdministered
      properties:
        name:
          type: string
          example: Distemper, Hepatitis, Parvovirus (DHP)
        dateAdministered:
          type: string
          format: date
          example: '2024-03-20'
        nextDueDate:
          type: string
          format: date
          example: '2025-03-20'
...
    WalkStatus:
      type: string
      enum:
        - scheduled
        - in_progress
        - completed
        - cancelled
      example: scheduled
    WalkEdit:
      type: object
      properties:
        scheduledDate:
          type: string
          format: date
          example: '2025-04-10'
        scheduledStartTime:
          type: string
          format: time
          example: '09:15:00Z'
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 240
          example: 60
        routeNotes:
          type: string
          description: Notes about the planned walking route
          example: Use the river path today.
...
    WalkCancel:
      type: object
      properties:
        reason:
          type: string
          description: Reason for cancelling the walk
          example: Dog owner had an emergency.
...
    WalkList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Walk'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Walk Updates ---------------------------------------------------------
...
    RecurringWalkStatus:
      type: string
      enum:
        - active
        - paused
        - cancelled
      example: active
...
    RecurringWalkList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RecurringWalk'
        pagination:
          $ref: '#/components/schemas/Pagination'
...